devices/requirements.txt
websockets==11.0.1
pyserial==3.5